% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-templates.R
\name{account-templates}
\alias{account-templates}
\title{Account Templates}
\description{
Functions for loading and applying account template structures.
Templates define standard chart of accounts for different entity types.
}
