% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{account_balance}
\alias{account_balance}
\title{Account Balance}
\usage{
account_balance(gc, account, as_of = Sys.Date())
}
\arguments{
\item{gc}{GnuCashDB object}

\item{account}{Account name, path, or GUID}

\item{as_of}{Date for balance calculation}
}
\value{
Numeric balance
}
\description{
Get the balance of a specific account as of a date.
}
