% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{account_transactions}
\alias{account_transactions}
\title{Get Account Transactions}
\usage{
account_transactions(gc, account, start = NULL, end = NULL)
}
\arguments{
\item{gc}{GnuCashDB object}

\item{account}{Account name or path}

\item{start}{Start date (optional)}

\item{end}{End date (optional)}
}
\value{
tibble of transactions
}
\description{
Retrieve transactions for a specific account.
}
