% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{account_tree}
\alias{account_tree}
\title{Get Account Tree}
\usage{
account_tree(gc)
}
\arguments{
\item{gc}{GnuCashDB object}
}
\value{
tibble with account hierarchy
}
\description{
Build hierarchical account structure with full paths.
}
