% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation.R
\name{aggregate_by_type}
\alias{aggregate_by_type}
\title{Aggregate by Account Type}
\usage{
aggregate_by_type(tb)
}
\arguments{
\item{tb}{Trial balance tibble}
}
\value{
Aggregated tibble by account type
}
\description{
Roll up trial balance to account type level.
}
