% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{annual_budget}
\alias{annual_budget}
\title{Create an Annual Budget}
\usage{
annual_budget(
  name,
  year = lubridate::year(Sys.Date()),
  description = NA_character_
)
}
\arguments{
\item{name}{Budget name}

\item{year}{Budget year (default: current year)}

\item{description}{Optional description}
}
\value{
A Budget object
}
\description{
Helper function to create a budget with a single annual period.
}
