% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled-transaction.R
\name{annual_scheduled}
\alias{annual_scheduled}
\title{Create an Annual Scheduled Transaction}
\usage{
annual_scheduled(
  name,
  amount,
  debit_account_guid,
  credit_account_guid,
  start_date,
  auto_create = FALSE
)
}
\arguments{
\item{name}{Transaction description}

\item{amount}{Transaction amount}

\item{debit_account_guid}{Account to debit}

\item{credit_account_guid}{Account to credit}

\item{start_date}{First occurrence}

\item{auto_create}{Auto-create transactions}
}
\value{
A ScheduledTransaction object
}
\description{
Helper function to create an annual recurring transaction.
}
