% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\name{approval_approve}
\alias{approval_approve}
\title{Approve a Request}
\usage{
approval_approve(approval_ptr, id, approver)
}
\arguments{
\item{approval_ptr}{External pointer from \code{\link{approval_open}}.}

\item{id}{Request ID to approve.}

\item{approver}{Identity of the approver.}
}
\description{
Approve a Request
}
