% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\name{approval_close}
\alias{approval_close}
\title{Close Approval Database}
\usage{
approval_close(approval_ptr)
}
\arguments{
\item{approval_ptr}{External pointer from \code{\link{approval_open}}.}
}
\description{
Close Approval Database
}
