% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\name{approval_get}
\alias{approval_get}
\title{Get Approval Request Details}
\usage{
approval_get(approval_ptr, id)
}
\arguments{
\item{approval_ptr}{External pointer from \code{\link{approval_open}}.}

\item{id}{Request ID.}
}
\value{
Named list with request details, or \code{NULL} if not found.
}
\description{
Get Approval Request Details
}
