% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\name{approval_open}
\alias{approval_open}
\title{Open Approval Database}
\usage{
approval_open(book_path)
}
\arguments{
\item{book_path}{Path to the GnuCash book file.}
}
\value{
External pointer to the approval database.
}
\description{
Opens or creates the approval database for a GnuCash book.
The database is stored at \code{<book_path>.approvals.db}.
}
