% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\name{approval_reject}
\alias{approval_reject}
\title{Reject a Request}
\usage{
approval_reject(approval_ptr, id, approver, reason)
}
\arguments{
\item{approval_ptr}{External pointer from \code{\link{approval_open}}.}

\item{id}{Request ID to reject.}

\item{approver}{Identity of the approver.}

\item{reason}{Reason for rejection.}
}
\description{
Reject a Request
}
