% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-logger.R
\name{audited_consolidation}
\alias{audited_consolidation}
\title{Audited Consolidation}
\usage{
audited_consolidation(collection, as_of = Sys.Date(), eliminate_ic = TRUE)
}
\arguments{
\item{collection}{BookCollection object}

\item{as_of}{Date for balances}

\item{eliminate_ic}{Apply IC eliminations}
}
\value{
Logged consolidated trial balance
}
\description{
Perform multi-book consolidation with audit logging.
}
