% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-logged.R
\name{audited_forecast}
\alias{audited_forecast}
\title{Create Audited Forecast}
\usage{
audited_forecast(lf, description = "Financial forecast")
}
\arguments{
\item{lf}{LazyForecast object}

\item{description}{Description of the forecast purpose}
}
\value{
Logged object containing LazyForecast
}
\description{
Wrap a LazyForecast with audit logging.
}
\examples{
\dontrun{
forecast <- forecast_expr(config) |>
  lf_monte_carlo(n = 1000)

audited <- audited_forecast(forecast, "Q1 2026 cash flow projection")
}

}
