% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-backup.R
\name{backup_summary}
\alias{backup_summary}
\title{Backup Info Summary}
\usage{
backup_summary(gc, backup_dir = NULL)
}
\arguments{
\item{gc}{GnuCashDB object or path}

\item{backup_dir}{Backup directory}
}
\value{
List with backup statistics
}
\description{
Get summary information about backups for a GnuCash file.
}
