% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book-collection.R
\name{book_collection}
\alias{book_collection}
\title{Create a Book Collection}
\usage{
book_collection(entity_config = NULL)
}
\arguments{
\item{entity_config}{Optional entity configuration}
}
\value{
A BookCollection object
}
\description{
Convenience function to create a BookCollection instance.
}
