% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{Budget}
\alias{Budget}
\title{Budget R6 Class}
\description{
Budget R6 Class

Budget R6 Class
}
\details{
R6 class representing a GnuCash budget.
Budgets have a defined period structure and amounts per account.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{guid}}{Budget GUID (read-only)}

\item{\code{name}}{Budget name}

\item{\code{description}}{Budget description}

\item{\code{num_periods}}{Number of periods}

\item{\code{recurrence_period_type}}{Period type}

\item{\code{recurrence_mult}}{Period multiplier}

\item{\code{recurrence_start}}{Start date}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Budget-new}{\code{Budget$new()}}
\item \href{#method-Budget-set_amount}{\code{Budget$set_amount()}}
\item \href{#method-Budget-get_amount}{\code{Budget$get_amount()}}
\item \href{#method-Budget-get_account_amounts}{\code{Budget$get_account_amounts()}}
\item \href{#method-Budget-get_account_total}{\code{Budget$get_account_total()}}
\item \href{#method-Budget-get_period_dates}{\code{Budget$get_period_dates()}}
\item \href{#method-Budget-period_labels}{\code{Budget$period_labels()}}
\item \href{#method-Budget-budgeted_accounts}{\code{Budget$budgeted_accounts()}}
\item \href{#method-Budget-amounts_as_tibble}{\code{Budget$amounts_as_tibble()}}
\item \href{#method-Budget-as_tibble}{\code{Budget$as_tibble()}}
\item \href{#method-Budget-print}{\code{Budget$print()}}
\item \href{#method-Budget-clone}{\code{Budget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-new"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-new}{}}}
\subsection{Method \code{new()}}{
Create a new Budget object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$new(
  guid = NULL,
  name,
  description = NA_character_,
  num_periods = 12L,
  recurrence_period_type = "month",
  recurrence_mult = 1L,
  recurrence_start = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{guid}}{Unique identifier (32-character hex)}

\item{\code{name}}{Budget name}

\item{\code{description}}{Optional description}

\item{\code{num_periods}}{Number of budget periods (12 for monthly, 4 for quarterly)}

\item{\code{recurrence_period_type}}{Period type (month, quarter, year)}

\item{\code{recurrence_mult}}{Multiplier for period}

\item{\code{recurrence_start}}{Start date for budget periods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-set_amount"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-set_amount}{}}}
\subsection{Method \code{set_amount()}}{
Set budget amount for an account and period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$set_amount(account_guid, period_num, amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{period_num}}{Period number (1-indexed)}

\item{\code{amount}}{Budget amount}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-get_amount"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-get_amount}{}}}
\subsection{Method \code{get_amount()}}{
Get budget amount for an account and period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$get_amount(account_guid, period_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{period_num}}{Period number (1-indexed)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-get_account_amounts"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-get_account_amounts}{}}}
\subsection{Method \code{get_account_amounts()}}{
Get all budget amounts for an account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$get_account_amounts(account_guid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-get_account_total"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-get_account_total}{}}}
\subsection{Method \code{get_account_total()}}{
Get total budget for an account across all periods
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$get_account_total(account_guid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-get_period_dates"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-get_period_dates}{}}}
\subsection{Method \code{get_period_dates()}}{
Get date range for a period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$get_period_dates(period_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period_num}}{Period number (1-indexed)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-period_labels"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-period_labels}{}}}
\subsection{Method \code{period_labels()}}{
Get all period labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$period_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-budgeted_accounts"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-budgeted_accounts}{}}}
\subsection{Method \code{budgeted_accounts()}}{
Get all accounts with budget amounts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$budgeted_accounts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-amounts_as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-amounts_as_tibble}{}}}
\subsection{Method \code{amounts_as_tibble()}}{
Get budget amounts as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$amounts_as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Convert to data frame row
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-print"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Budget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Budget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
