% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{budget_from_df}
\alias{budget_from_df}
\title{Convert Data Frame to Budget Object}
\usage{
budget_from_df(budget_row, amounts_df = NULL)
}
\arguments{
\item{budget_row}{Single row from budgets table}

\item{amounts_df}{Budget amounts data frame}
}
\value{
A Budget object
}
\description{
Create a Budget from budget and budget_amounts data frames.
}
