% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{BudgetComparison}
\alias{BudgetComparison}
\title{BudgetComparison R6 Class}
\description{
BudgetComparison R6 Class

BudgetComparison R6 Class
}
\details{
R6 class for comparing budget to actuals.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{budget}}{The budget being compared}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BudgetComparison-new}{\code{BudgetComparison$new()}}
\item \href{#method-BudgetComparison-set_actual}{\code{BudgetComparison$set_actual()}}
\item \href{#method-BudgetComparison-get_actual}{\code{BudgetComparison$get_actual()}}
\item \href{#method-BudgetComparison-variance}{\code{BudgetComparison$variance()}}
\item \href{#method-BudgetComparison-variance_pct}{\code{BudgetComparison$variance_pct()}}
\item \href{#method-BudgetComparison-account_summary}{\code{BudgetComparison$account_summary()}}
\item \href{#method-BudgetComparison-as_tibble}{\code{BudgetComparison$as_tibble()}}
\item \href{#method-BudgetComparison-print}{\code{BudgetComparison$print()}}
\item \href{#method-BudgetComparison-clone}{\code{BudgetComparison$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-new"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-new}{}}}
\subsection{Method \code{new()}}{
Create a new BudgetComparison
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$new(budget)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{budget}}{A Budget object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-set_actual"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-set_actual}{}}}
\subsection{Method \code{set_actual()}}{
Set actual amount for an account and period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$set_actual(account_guid, period_num, amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{period_num}}{Period number}

\item{\code{amount}}{Actual amount}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-get_actual"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-get_actual}{}}}
\subsection{Method \code{get_actual()}}{
Get actual amount for an account and period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$get_actual(account_guid, period_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{period_num}}{Period number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-variance"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-variance}{}}}
\subsection{Method \code{variance()}}{
Get variance (actual - budget) for an account and period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$variance(account_guid, period_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{period_num}}{Period number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-variance_pct"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-variance_pct}{}}}
\subsection{Method \code{variance_pct()}}{
Get variance percentage for an account and period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$variance_pct(account_guid, period_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{period_num}}{Period number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-account_summary"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-account_summary}{}}}
\subsection{Method \code{account_summary()}}{
Get comparison summary for an account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$account_summary(account_guid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Get full comparison as tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-print"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BudgetComparison-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BudgetComparison-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BudgetComparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
