% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation.R
\name{build_consolidated_trial_balance}
\alias{build_consolidated_trial_balance}
\title{Build Consolidated Trial Balance}
\usage{
build_consolidated_trial_balance(books, ic_rules = list(), as_of = Sys.Date())
}
\arguments{
\item{books}{List of GnuCashDB objects}

\item{ic_rules}{Intercompany rules}

\item{as_of}{Date for balances}
}
\value{
Consolidated trial balance tibble
}
\description{
Create fully consolidated trial balance with all adjustments.
}
