% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{combine_results}
\alias{combine_results}
\title{Combine Results}
\usage{
combine_results(...)
}
\arguments{
\item{...}{Result objects to combine}
}
\value{
Result with list of values or first error
}
\description{
Combine multiple results, failing if any fail.
}
