% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodity.R
\name{Commodity}
\alias{Commodity}
\title{Commodity R6 Class}
\description{
Commodity R6 Class

Commodity R6 Class
}
\details{
R6 class representing a GnuCash commodity (currency, stock, mutual fund, etc.).
Commodities are identified by their namespace and mnemonic (e.g., "CURRENCY:USD").
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{guid}}{Commodity GUID (read-only)}

\item{\code{namespace}}{Commodity namespace}

\item{\code{mnemonic}}{Commodity mnemonic}

\item{\code{fullname}}{Full descriptive name}

\item{\code{cusip}}{CUSIP/ISIN identifier}

\item{\code{fraction}}{Smallest unit (100 = cents)}

\item{\code{quote_flag}}{Whether to fetch online quotes}

\item{\code{quote_source}}{Source for online quotes}

\item{\code{quote_tz}}{Timezone for quotes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Commodity-new}{\code{Commodity$new()}}
\item \href{#method-Commodity-identifier}{\code{Commodity$identifier()}}
\item \href{#method-Commodity-is_currency}{\code{Commodity$is_currency()}}
\item \href{#method-Commodity-is_security}{\code{Commodity$is_security()}}
\item \href{#method-Commodity-decimal_places}{\code{Commodity$decimal_places()}}
\item \href{#method-Commodity-to_smallest_unit}{\code{Commodity$to_smallest_unit()}}
\item \href{#method-Commodity-from_smallest_unit}{\code{Commodity$from_smallest_unit()}}
\item \href{#method-Commodity-format_amount}{\code{Commodity$format_amount()}}
\item \href{#method-Commodity-as_tibble}{\code{Commodity$as_tibble()}}
\item \href{#method-Commodity-print}{\code{Commodity$print()}}
\item \href{#method-Commodity-clone}{\code{Commodity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-new"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-new}{}}}
\subsection{Method \code{new()}}{
Create a new Commodity object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$new(
  guid = NULL,
  namespace = "CURRENCY",
  mnemonic,
  fullname = NA_character_,
  cusip = NA_character_,
  fraction = 100L,
  quote_flag = FALSE,
  quote_source = NA_character_,
  quote_tz = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{guid}}{Unique identifier (32-character hex)}

\item{\code{namespace}}{Commodity namespace (CURRENCY, NASDAQ, NYSE, etc.)}

\item{\code{mnemonic}}{Short identifier (USD, AAPL, etc.)}

\item{\code{fullname}}{Full descriptive name}

\item{\code{cusip}}{CUSIP/ISIN identifier for securities}

\item{\code{fraction}}{Smallest unit (100 for USD = cents)}

\item{\code{quote_flag}}{Whether to fetch online quotes}

\item{\code{quote_source}}{Source for online quotes}

\item{\code{quote_tz}}{Timezone for quotes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-identifier"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-identifier}{}}}
\subsection{Method \code{identifier()}}{
Get commodity identifier (namespace:mnemonic)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$identifier()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-is_currency"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-is_currency}{}}}
\subsection{Method \code{is_currency()}}{
Check if this is a currency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$is_currency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-is_security"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-is_security}{}}}
\subsection{Method \code{is_security()}}{
Check if this is a security (stock, mutual fund, etc.)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$is_security()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-decimal_places"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-decimal_places}{}}}
\subsection{Method \code{decimal_places()}}{
Get number of decimal places
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$decimal_places()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-to_smallest_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-to_smallest_unit}{}}}
\subsection{Method \code{to_smallest_unit()}}{
Convert amount to smallest unit (e.g., dollars to cents)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$to_smallest_unit(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Numeric amount to convert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-from_smallest_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-from_smallest_unit}{}}}
\subsection{Method \code{from_smallest_unit()}}{
Convert from smallest unit to decimal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$from_smallest_unit(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Integer amount in smallest units}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-format_amount"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-format_amount}{}}}
\subsection{Method \code{format_amount()}}{
Format amount according to commodity precision
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$format_amount(amount, include_symbol = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Numeric amount}

\item{\code{include_symbol}}{Include currency symbol}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Convert to data frame row
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-print"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commodity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Commodity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commodity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
