% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-templates.R
\name{compare_to_template}
\alias{compare_to_template}
\title{Compare Template to Existing Accounts}
\usage{
compare_to_template(gc, template)
}
\arguments{
\item{gc}{A GnuCashDB object}

\item{template}{Template list (from load_template) or template name}
}
\value{
A list with missing and extra accounts
}
\description{
Compare a template with existing accounts in a GnuCash file.
Identifies missing and extra accounts.
}
