% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation.R
\name{consolidation_summary}
\alias{consolidation_summary}
\title{Consolidation Summary Report}
\usage{
consolidation_summary(books, ic_rules = list(), as_of = Sys.Date())
}
\arguments{
\item{books}{List of GnuCashDB objects}

\item{ic_rules}{Intercompany rules}

\item{as_of}{Date for balances}
}
\value{
List with consolidation details
}
\description{
Generate a summary of the consolidation process.
}
