% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercompany.R
\name{create_standard_ic_rules}
\alias{create_standard_ic_rules}
\title{Create Standard IC Rule Set}
\usage{
create_standard_ic_rules(
  parent_name = "inc",
  subsidiary_names = c("products", "services", "software", "operations"),
  parent_ar_pattern = "Assets:Due from \%s",
  sub_ap_pattern = "Liabilities:Due to Parent",
  mgmt_fee_income = "Income:Management Fees:\%s",
  mgmt_fee_expense = "Expenses:Management Fees"
)
}
\arguments{
\item{parent_name}{Name of parent company book}

\item{subsidiary_names}{Character vector of subsidiary book names}

\item{parent_ar_pattern}{Pattern for parent's receivable accounts}

\item{sub_ap_pattern}{Pattern for subsidiary's payable accounts}

\item{mgmt_fee_income}{Pattern for management fee income}

\item{mgmt_fee_expense}{Pattern for management fee expense}
}
\value{
List of IC rules
}
\description{
Generate standard intercompany elimination rules for a
parent company with multiple subsidiaries.
}
