% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-account.R
\name{delete_account}
\alias{delete_account}
\title{Delete Account}
\usage{
delete_account(gc, account_guid, force = FALSE)
}
\arguments{
\item{gc}{GnuCashDB object}

\item{account_guid}{Account GUID to delete}

\item{force}{Force delete even with children (will reparent children to grandparent)}
}
\value{
Result with deleted GUID or error
}
\description{
Delete an account (must have no transactions and no children).
}
