% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slots.R
\name{delete_slot}
\alias{delete_slot}
\title{Delete Slot}
\usage{
delete_slot(book_ptr, obj_guid, name)
}
\arguments{
\item{book_ptr}{External pointer from \code{gc_open(read_only = FALSE)}}

\item{obj_guid}{Entity GUID}

\item{name}{Slot name to delete}
}
\value{
Invisible NULL
}
\description{
Remove a slot from an entity.
}
