% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-transaction.R
\name{delete_transaction}
\alias{delete_transaction}
\title{Delete Transaction}
\usage{
delete_transaction(gc, tx_guid)
}
\arguments{
\item{gc}{GnuCashDB object}

\item{tx_guid}{Transaction GUID to delete}
}
\value{
Result with deleted GUID or error
}
\description{
Delete a transaction and all its splits.
}
