% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnucash-xml.R
\name{detect_gnucash_format}
\alias{detect_gnucash_format}
\title{Detect GnuCash File Format}
\usage{
detect_gnucash_format(path)
}
\arguments{
\item{path}{Path to GnuCash file}
}
\value{
Character: "sqlite", "xml-gz", "xml", or "unknown"
}
\description{
Determine if a GnuCash file is SQLite or XML format.
}
