% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-ofx.R
\name{detect_ofx_version}
\alias{detect_ofx_version}
\title{Detect OFX Version}
\usage{
detect_ofx_version(path)
}
\arguments{
\item{path}{Path to the OFX/QFX file}
}
\value{
Character string: "1" for SGML, "2" for XML, "unknown" otherwise
}
\description{
Determine whether an OFX file is version 1.x (SGML-based) or
version 2.x (XML-based).
}
\examples{
\dontrun{
version <- detect_ofx_version("statement.ofx")
if (version == "1") {
  message("OFX 1.x SGML format detected")
}
}
}
