% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_ofx_version_cpp}
\alias{detect_ofx_version_cpp}
\title{Detect OFX File Version}
\usage{
detect_ofx_version_cpp(content)
}
\arguments{
\item{content}{String content of the OFX file}
}
\value{
String: "1" for SGML, "2" for XML, "unknown" otherwise
}
\description{
Determine whether an OFX file is version 1.x (SGML) or 2.x (XML).
}
