% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{err}
\alias{err}
\title{Create Error Result}
\usage{
err(error)
}
\arguments{
\item{error}{Error message or condition}
}
\value{
A result object with error status
}
\description{
Wrap an error condition.
}
