% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-result.R
\name{extract_mc_summary}
\alias{extract_mc_summary}
\title{Extract Summary from Result}
\usage{
extract_mc_summary(result)
}
\arguments{
\item{result}{Result from safe_collect of Monte Carlo forecast}
}
\value{
Result with summary or error
}
\description{
Safely extract summary statistics from a Monte Carlo result.
}
