% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_ofx_account_info}
\alias{extract_ofx_account_info}
\title{Extract OFX Account Info}
\usage{
extract_ofx_account_info(content)
}
\arguments{
\item{content}{String content of the OFX file}
}
\value{
List with account_id, bank_id, account_type, org_name
}
\description{
Extract account information from OFX content (ACCTID, BANKID, etc.)
}
