% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-templates.R
\name{flatten_template}
\alias{flatten_template}
\title{Flatten Account Template to Data Frame}
\usage{
flatten_template(template, include_guids = TRUE)
}
\arguments{
\item{template}{Template list (from load_template) or template name}

\item{include_guids}{Generate GUIDs for each account}
}
\value{
A tibble with account data
}
\description{
Convert a hierarchical account template to a flat data frame.
Useful for importing into GnuCash.
}
\examples{
\dontrun{
accounts <- flatten_template("small-business")
}
}
