% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fraction_to_double}
\alias{fraction_to_double}
\title{Convert Fractions to Double}
\usage{
fraction_to_double(numerator, denominator)
}
\arguments{
\item{numerator}{Integer vector of numerators (value_num)}

\item{denominator}{Integer vector of denominators (value_denom)}
}
\value{
NumericVector of double values
}
\description{
Vectorized conversion of GnuCash fraction representation to double.
This is a hot path in balance calculations, hence Rcpp implementation.
}
