% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-expr.R
\name{from_book}
\alias{from_book}
\title{Create Forecast from GnuCash Book}
\usage{
from_book(gc)
}
\arguments{
\item{gc}{GnuCashDB object}
}
\value{
LazyForecast with GnuCashDB source
}
\description{
Initialize a lazy forecast from a GnuCash database.
}
\examples{
\dontrun{
gc <- read_gnucash("path/to/file.gnucash")
forecast <- from_book(gc) |>
  grow(rate = 0.05, months = 12)
}

}
