% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-expr.R
\name{from_collection}
\alias{from_collection}
\title{Create Forecast from Book Collection}
\usage{
from_collection(collection)
}
\arguments{
\item{collection}{BookCollection object}
}
\value{
LazyForecast with BookCollection source
}
\description{
Initialize a lazy forecast from a multi-book collection.
}
\examples{
\dontrun{
collection <- book_collection()
collection$add_book("inc", "path/to/inc.gnucash")
forecast <- from_collection(collection) |>
  grow(rate = 0.05, months = 12)
}

}
