% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_agent_state_set}
\alias{gc_agent_state_set}
\title{Set Agent State Value (C++ Backend)}
\usage{
gc_agent_state_set(state_ptr, key, value)
}
\arguments{
\item{state_ptr}{External pointer from gc_agent_state_open()}

\item{key}{State key}

\item{value}{State value}
}
\description{
Set Agent State Value (C++ Backend)
}
