% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_approval_approve}
\alias{gc_approval_approve}
\title{Approve a Pending Request (C++ Backend)}
\usage{
gc_approval_approve(approval_ptr, id, approver)
}
\arguments{
\item{approval_ptr}{External pointer from gc_approval_open()}

\item{id}{Request ID to approve}

\item{approver}{Identity of the approver}
}
\description{
Approve a Pending Request (C++ Backend)
}
