% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_approval_close}
\alias{gc_approval_close}
\title{Close Approval Database (C++ Backend)}
\usage{
gc_approval_close(approval_ptr)
}
\arguments{
\item{approval_ptr}{External pointer from gc_approval_open()}
}
\description{
Close Approval Database (C++ Backend)
}
