% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_approval_get}
\alias{gc_approval_get}
\title{Get a Specific Approval Request (C++ Backend)}
\usage{
gc_approval_get(approval_ptr, id)
}
\arguments{
\item{approval_ptr}{External pointer from gc_approval_open()}

\item{id}{Request ID}
}
\value{
Named list or NULL if not found
}
\description{
Get a Specific Approval Request (C++ Backend)
}
