% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_approval_open}
\alias{gc_approval_open}
\title{Open Approval Database (C++ Backend)}
\usage{
gc_approval_open(book_path)
}
\arguments{
\item{book_path}{Path to GnuCash file (approval DB at \if{html}{\out{<book>}}.approvals.db)}
}
\value{
External pointer to ApprovalDB
}
\description{
Open Approval Database (C++ Backend)
}
