% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_approval_pending}
\alias{gc_approval_pending}
\title{List Pending Approval Requests (C++ Backend)}
\usage{
gc_approval_pending(approval_ptr, limit = 50L)
}
\arguments{
\item{approval_ptr}{External pointer from gc_approval_open()}

\item{limit}{Maximum number of requests to return}
}
\value{
Data frame of pending requests
}
\description{
List Pending Approval Requests (C++ Backend)
}
