% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_audit_close}
\alias{gc_audit_close}
\title{Close Audit Database (C++ Backend)}
\usage{
gc_audit_close(audit_ptr)
}
\arguments{
\item{audit_ptr}{External pointer from gc_audit_open()}
}
\description{
Close Audit Database (C++ Backend)
}
