% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_close}
\alias{gc_close}
\title{Close a GnuCash Database Handle}
\usage{
gc_close(book_ptr)
}
\arguments{
\item{book_ptr}{External pointer from gc_open()}
}
\description{
Close a GnuCash Database Handle
}
