% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_delete_slot}
\alias{gc_delete_slot}
\title{Delete Slot (C++ Backend)}
\usage{
gc_delete_slot(book_ptr, obj_guid, name)
}
\arguments{
\item{book_ptr}{External pointer from gc_open() (must be read-write)}

\item{obj_guid}{Entity GUID}

\item{name}{Slot name}
}
\description{
Delete Slot (C++ Backend)
}
