% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_delete_transaction}
\alias{gc_delete_transaction}
\title{Delete Transaction (C++ Backend)}
\usage{
gc_delete_transaction(book_ptr, guid)
}
\arguments{
\item{book_ptr}{External pointer from gc_open() (must be read-write)}

\item{guid}{Transaction GUID}
}
\description{
Delete Transaction (C++ Backend)
}
