% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_get_accounts}
\alias{gc_get_accounts}
\title{Get All Accounts (C++ Backend)}
\usage{
gc_get_accounts(book_ptr)
}
\arguments{
\item{book_ptr}{External pointer from gc_open()}
}
\value{
Data frame of accounts
}
\description{
Get All Accounts (C++ Backend)
}
