% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_get_all_slots}
\alias{gc_get_all_slots}
\title{Get All Slots for Entity (C++ Backend)}
\usage{
gc_get_all_slots(book_ptr, obj_guid)
}
\arguments{
\item{book_ptr}{External pointer from gc_open()}

\item{obj_guid}{Entity GUID}
}
\value{
Data frame of all slots
}
\description{
Get All Slots for Entity (C++ Backend)
}
