% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_get_splits}
\alias{gc_get_splits}
\title{Get Splits for Account (C++ Backend)}
\usage{
gc_get_splits(book_ptr, account_guid)
}
\arguments{
\item{book_ptr}{External pointer from gc_open()}

\item{account_guid}{Account GUID}
}
\value{
Data frame of splits ordered by post_date
}
\description{
Get Splits for Account (C++ Backend)
}
