% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_info}
\alias{gc_info}
\title{Get Book Metadata}
\usage{
gc_info(book_ptr)
}
\arguments{
\item{book_ptr}{External pointer from gc_open()}
}
\value{
Named list with book_guid, root_account_guid, default_currency, read_only
}
\description{
Get Book Metadata
}
