% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_rate_limiter_create}
\alias{gc_rate_limiter_create}
\title{Create Rate Limiter (C++ Backend)}
\usage{
gc_rate_limiter_create()
}
\value{
External pointer to a RateLimiter
}
\description{
Create Rate Limiter (C++ Backend)
}
