% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_set_slot}
\alias{gc_set_slot}
\title{Set String Slot Value (C++ Backend)}
\usage{
gc_set_slot(book_ptr, obj_guid, name, value)
}
\arguments{
\item{book_ptr}{External pointer from gc_open() (must be read-write)}

\item{obj_guid}{Entity GUID}

\item{name}{Slot name}

\item{value}{Slot value}
}
\description{
Set a string-typed slot on an entity. Creates or updates.
}
