% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_system_username}
\alias{gc_system_username}
\title{Get System Username (C++ Backend)}
\usage{
gc_system_username()
}
\value{
System username string
}
\description{
Get System Username (C++ Backend)
}
