% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_void_transaction}
\alias{gc_void_transaction}
\title{Void Transaction (C++ Backend)}
\usage{
gc_void_transaction(book_ptr, guid, reason)
}
\arguments{
\item{book_ptr}{External pointer from gc_open() (must be read-write)}

\item{guid}{Transaction GUID}

\item{reason}{Void reason}
}
\description{
Void Transaction (C++ Backend)
}
