% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_guid}
\alias{generate_guid}
\title{Generate GnuCash-Format GUID}
\usage{
generate_guid()
}
\value{
Character string with 32-character hex GUID
}
\description{
Generate a random 32-character hex GUID in GnuCash format.
GnuCash uses lowercase hex without dashes.
}
