% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_guids}
\alias{generate_guids}
\title{Generate Multiple GUIDs}
\usage{
generate_guids(n)
}
\arguments{
\item{n}{Number of GUIDs to generate}
}
\value{
CharacterVector of GUIDs
}
\description{
Vectorized GUID generation for batch operations.
}
