% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnucash-xml.R
\name{gnucash-xml}
\alias{gnucash-xml}
\title{GnuCash XML Parser}
\description{
Functions for reading GnuCash XML format files (compressed or uncompressed).
GnuCash XML files are typically gzip-compressed by default.
}
