% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercompany.R
\name{identify_uneliminated_ic}
\alias{identify_uneliminated_ic}
\title{Identify Uneliminated IC Transactions}
\usage{
identify_uneliminated_ic(
  collection,
  ic_keywords = c("due from", "due to", "intercompany", "affiliate", "management fee",
    "license fee", "inter-co")
)
}
\arguments{
\item{collection}{BookCollection object}

\item{ic_keywords}{Keywords that indicate intercompany accounts}
}
\value{
tibble of potentially uneliminated IC transactions
}
\description{
Find intercompany transactions that should have been eliminated
but weren't matched by existing rules.
}
