% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-ofx.R
\name{import-ofx}
\alias{import-ofx}
\title{OFX/QFX Import Functions}
\description{
Functions for importing bank statements in OFX (Open Financial Exchange)
and QFX (Quicken Financial Exchange) formats. Supports both OFX 1.x (SGML)
and OFX 2.x (XML) file formats.
}
