% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-income-statement.R
\name{income_statement}
\alias{income_statement}
\title{Generate Income Statement}
\usage{
income_statement(data)
}
\arguments{
\item{data}{Activity tibble from income_statement_activity() or trial_balance()}
}
\value{
tibble with income statement structure
}
\description{
Create income statement from activity data or trial balance.
}
