% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-income-statement.R
\name{income_statement_activity}
\alias{income_statement_activity}
\title{Get Income Statement Activity}
\usage{
income_statement_activity(gc, start_date, end_date = Sys.Date())
}
\arguments{
\item{gc}{GnuCashDB object}

\item{start_date}{Start of period}

\item{end_date}{End of period}
}
\value{
tibble with income/expense activity
}
\description{
Extract income and expense activity for a specific period.
}
