% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-csv.R
\name{is_csv_import}
\alias{is_csv_import}
\title{Check if Object is a gnucashr CSV Import}
\usage{
is_csv_import(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical TRUE if x is a gnucashr_csv_import
}
\description{
Check if Object is a gnucashr CSV Import
}
\examples{
\dontrun{
import <- import_paypal_csv("paypal_export.csv")
is_csv_import(import)  # TRUE
}
}
