% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-lazy.R
\name{lazy_forecast}
\alias{lazy_forecast}
\title{Create a new LazyForecast}
\usage{
lazy_forecast(source = NULL)
}
\arguments{
\item{source}{Data source (config, tibble, GnuCashDB, BookCollection)}
}
\value{
LazyForecast object
}
\description{
Factory function for creating LazyForecast objects.
}
