% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-expr.R
\name{lf_filter}
\alias{lf_filter}
\title{Filter Forecast Data (Pipe-Friendly)}
\usage{
lf_filter(lf, ...)
}
\arguments{
\item{lf}{LazyForecast object}

\item{...}{Filter conditions}
}
\value{
LazyForecast (for chaining)
}
\description{
Add filter operation to a lazy forecast.
}
