% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-expr.R
\name{lf_sensitivity}
\alias{lf_sensitivity}
\title{Sensitivity Analysis (Pipe-Friendly)}
\usage{
lf_sensitivity(
  lf,
  growth_range = seq(-0.02, 0.08, by = 0.01),
  expense_range = seq(0.5, 0.9, by = 0.05),
  months = 12
)
}
\arguments{
\item{lf}{LazyForecast object}

\item{growth_range}{Vector of growth rates to test}

\item{expense_range}{Vector of expense ratios to test}

\item{months}{Number of months to project}
}
\value{
LazyForecast (for chaining)
}
\description{
Add sensitivity grid analysis to a lazy forecast.
}
\examples{
\dontrun{
forecast <- forecast_expr() |>
  lf_sensitivity(
    growth_range = seq(-0.02, 0.08, by = 0.01),
    expense_range = seq(0.5, 0.9, by = 0.05)
  )
result <- forecast$collect()
# result$outcomes is a matrix of results
}

}
