% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-backup.R
\name{list_backups}
\alias{list_backups}
\title{List Available Backups}
\usage{
list_backups(gc, backup_dir = NULL)
}
\arguments{
\item{gc}{GnuCashDB object or path to GnuCash file}

\item{backup_dir}{Backup directory (default: auto-detect)}
}
\value{
tibble of available backups
}
\description{
List all available backups for a GnuCash file.
}
