% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-templates.R
\name{list_templates}
\alias{list_templates}
\title{List Available Account Templates}
\usage{
list_templates()
}
\value{
A tibble with template names and descriptions
}
\description{
Get a list of built-in account templates.
}
\examples{
\dontrun{
list_templates()
}
}
