% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-logger.R
\name{log_append}
\alias{log_append}
\title{Append to Log}
\usage{
log_append(x, entry)
}
\arguments{
\item{x}{A logged object}

\item{entry}{New log entry string}
}
\value{
Updated logged object
}
\description{
Add a new entry to the log.
}
