% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-logger.R
\name{log_map}
\alias{log_map}
\title{Map Over Logged Value}
\usage{
log_map(x, fn)
}
\arguments{
\item{x}{A logged object}

\item{fn}{Function to apply to the value}
}
\value{
Logged object with transformed value
}
\description{
Apply a function to the value without modifying the log.
}
