% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-logger.R
\name{logged_log}
\alias{logged_log}
\title{Get Log from Logged Result}
\usage{
logged_log(x)
}
\arguments{
\item{x}{A logged object}
}
\value{
Character vector of log entries
}
\description{
Extract the audit log from a logged operation.
}
