% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-logger.R
\name{logged_value}
\alias{logged_value}
\title{Get Value from Logged Result}
\usage{
logged_value(x)
}
\arguments{
\item{x}{A logged object}
}
\value{
The unwrapped value
}
\description{
Extract the value from a logged operation.
}
