% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lot.R
\name{LotManager}
\alias{LotManager}
\title{LotManager R6 Class}
\description{
LotManager R6 Class

LotManager R6 Class
}
\details{
R6 class for managing lots for an account.
Provides FIFO, LIFO, and specific identification methods.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{method}}{Cost basis method}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LotManager-new}{\code{LotManager$new()}}
\item \href{#method-LotManager-add_lot}{\code{LotManager$add_lot()}}
\item \href{#method-LotManager-lots}{\code{LotManager$lots()}}
\item \href{#method-LotManager-open_lots_ordered}{\code{LotManager$open_lots_ordered()}}
\item \href{#method-LotManager-total_quantity}{\code{LotManager$total_quantity()}}
\item \href{#method-LotManager-total_cost_basis}{\code{LotManager$total_cost_basis()}}
\item \href{#method-LotManager-avg_cost}{\code{LotManager$avg_cost()}}
\item \href{#method-LotManager-unrealized_gain}{\code{LotManager$unrealized_gain()}}
\item \href{#method-LotManager-select_lots_for_sale}{\code{LotManager$select_lots_for_sale()}}
\item \href{#method-LotManager-as_tibble}{\code{LotManager$as_tibble()}}
\item \href{#method-LotManager-print}{\code{LotManager$print()}}
\item \href{#method-LotManager-clone}{\code{LotManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new LotManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$new(account_guid, method = "FIFO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_guid}}{Account GUID}

\item{\code{method}}{Cost basis method (FIFO, LIFO, SPECIFIC)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-add_lot"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-add_lot}{}}}
\subsection{Method \code{add_lot()}}{
Add a lot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$add_lot(lot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lot}}{A Lot object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-lots"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-lots}{}}}
\subsection{Method \code{lots()}}{
Get all lots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$lots(include_closed = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_closed}}{Include closed lots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-open_lots_ordered"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-open_lots_ordered}{}}}
\subsection{Method \code{open_lots_ordered()}}{
Get open lots sorted by method order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$open_lots_ordered()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-total_quantity"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-total_quantity}{}}}
\subsection{Method \code{total_quantity()}}{
Calculate total quantity across all open lots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$total_quantity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-total_cost_basis"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-total_cost_basis}{}}}
\subsection{Method \code{total_cost_basis()}}{
Calculate total cost basis across all open lots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$total_cost_basis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-avg_cost"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-avg_cost}{}}}
\subsection{Method \code{avg_cost()}}{
Calculate weighted average cost
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$avg_cost()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-unrealized_gain"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-unrealized_gain}{}}}
\subsection{Method \code{unrealized_gain()}}{
Calculate total unrealized gain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$unrealized_gain(current_price)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{current_price}}{Current market price}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-select_lots_for_sale"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-select_lots_for_sale}{}}}
\subsection{Method \code{select_lots_for_sale()}}{
Get lots to use for selling (based on method)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$select_lots_for_sale(quantity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantity}}{Quantity to sell}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Get all lots as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-print"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LotManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LotManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LotManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
