% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lot.R
\name{lots_from_df}
\alias{lots_from_df}
\title{Convert Data Frame to Lot Objects}
\usage{
lots_from_df(df)
}
\arguments{
\item{df}{Data frame with lot columns}
}
\value{
Named list of Lot objects (keyed by guid)
}
\description{
Convert a tibble of lot data to a list of Lot objects.
}
