% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-logged.R
\name{map_logged_forecast}
\alias{map_logged_forecast}
\title{Map Over Logged Forecast}
\usage{
map_logged_forecast(logged_lf, fn)
}
\arguments{
\item{logged_lf}{Logged object}

\item{fn}{Function to apply}
}
\value{
Logged object with transformed value
}
\description{
Apply a function to a logged forecast value without modifying the log.
}
