% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{monthly_activity}
\alias{monthly_activity}
\title{Monthly Activity Summary}
\usage{
monthly_activity(gc, account = NULL, start = NULL, end = NULL)
}
\arguments{
\item{gc}{GnuCashDB object}

\item{account}{Account name or path (NULL for all accounts)}

\item{start}{Start date}

\item{end}{End date}
}
\value{
tibble with monthly summaries
}
\description{
Summarize account activity by month.
}
