% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{monthly_budget}
\alias{monthly_budget}
\title{Create a Monthly Budget}
\usage{
monthly_budget(
  name,
  year = lubridate::year(Sys.Date()),
  description = NA_character_
)
}
\arguments{
\item{name}{Budget name}

\item{year}{Budget year (default: current year)}

\item{description}{Optional description}
}
\value{
A Budget object
}
\description{
Helper function to create a budget with monthly periods.
}
\examples{
\dontrun{
budget <- monthly_budget("2024 Operating Budget", 2024)
}
}
