% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled-transaction.R
\name{monthly_scheduled}
\alias{monthly_scheduled}
\title{Create a Monthly Scheduled Transaction}
\usage{
monthly_scheduled(
  name,
  amount,
  debit_account_guid,
  credit_account_guid,
  start_date,
  day_of_month = NULL,
  auto_create = FALSE
)
}
\arguments{
\item{name}{Transaction description}

\item{amount}{Transaction amount}

\item{debit_account_guid}{Account to debit}

\item{credit_account_guid}{Account to credit}

\item{start_date}{First occurrence}

\item{day_of_month}{Day of month (1-31, or "last")}

\item{auto_create}{Auto-create transactions}
}
\value{
A ScheduledTransaction object
}
\description{
Helper function to create a monthly recurring transaction.
}
