% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{ok}
\alias{ok}
\title{Create Success Result}
\usage{
ok(value)
}
\arguments{
\item{value}{The successful result value}
}
\value{
A result object with ok status
}
\description{
Wrap a successful value.
}
