% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnucash-xml.R
\name{parse_gnucash_xml}
\alias{parse_gnucash_xml}
\title{Parse GnuCash XML File}
\usage{
parse_gnucash_xml(path)
}
\arguments{
\item{path}{Path to GnuCash XML file}
}
\value{
List containing tibbles: accounts, transactions, splits, commodities, prices
}
\description{
Read and parse a GnuCash XML file into tibbles.
Handles both compressed (.gnucash) and uncompressed (.gnucash.xml) files.
}
