% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-csv.R
\name{preview_csv_mapping}
\alias{preview_csv_mapping}
\title{Preview CSV Column Mapping}
\usage{
preview_csv_mapping(path, import_fn)
}
\arguments{
\item{path}{Path to CSV file}

\item{import_fn}{Import function to use (e.g., import_paypal_csv)}
}
\value{
A tibble showing column mappings and sample values
}
\description{
Debug helper to show how columns from a CSV file would be mapped
by one of the import functions.
}
\examples{
\dontrun{
# Preview how PayPal columns will be mapped
preview_csv_mapping("paypal.csv", import_paypal_csv)
}
}
