% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{Price}
\alias{Price}
\title{Price R6 Class}
\description{
Price R6 Class

Price R6 Class
}
\details{
R6 class representing a price/exchange rate between two commodities.
Used for converting between currencies and tracking security prices.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{guid}}{Price GUID (read-only)}

\item{\code{commodity_guid}}{Commodity being priced}

\item{\code{currency_guid}}{Currency for pricing}

\item{\code{date}}{Price date/time}

\item{\code{source}}{Price source}

\item{\code{type}}{Price type (last, bid, ask, nav)}

\item{\code{value_num}}{Price numerator}

\item{\code{value_denom}}{Price denominator}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Price-new}{\code{Price$new()}}
\item \href{#method-Price-value}{\code{Price$value()}}
\item \href{#method-Price-set_value}{\code{Price$set_value()}}
\item \href{#method-Price-convert_to_currency}{\code{Price$convert_to_currency()}}
\item \href{#method-Price-convert_to_commodity}{\code{Price$convert_to_commodity()}}
\item \href{#method-Price-as_tibble}{\code{Price$as_tibble()}}
\item \href{#method-Price-print}{\code{Price$print()}}
\item \href{#method-Price-clone}{\code{Price$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-new"></a>}}
\if{latex}{\out{\hypertarget{method-Price-new}{}}}
\subsection{Method \code{new()}}{
Create a new Price object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$new(
  guid = NULL,
  commodity_guid,
  currency_guid,
  date = Sys.time(),
  source = "user:price",
  type = "last",
  value_num,
  value_denom = 100L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{guid}}{Unique identifier (32-character hex)}

\item{\code{commodity_guid}}{GUID of the commodity being priced}

\item{\code{currency_guid}}{GUID of the currency used for pricing}

\item{\code{date}}{Date/time of the price}

\item{\code{source}}{Source of the price (user, yahoo, etc.)}

\item{\code{type}}{Price type (last, bid, ask, nav, etc.)}

\item{\code{value_num}}{Numerator of price fraction}

\item{\code{value_denom}}{Denominator of price fraction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-value"></a>}}
\if{latex}{\out{\hypertarget{method-Price-value}{}}}
\subsection{Method \code{value()}}{
Get the price as a decimal value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Price-set_value}{}}}
\subsection{Method \code{set_value()}}{
Set the price from a decimal value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$set_value(new_value, denom = 100L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_value}}{New price value}

\item{\code{denom}}{Denominator for storage precision}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-convert_to_currency"></a>}}
\if{latex}{\out{\hypertarget{method-Price-convert_to_currency}{}}}
\subsection{Method \code{convert_to_currency()}}{
Convert an amount from commodity to currency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$convert_to_currency(commodity_amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commodity_amount}}{Amount in commodity units}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-convert_to_commodity"></a>}}
\if{latex}{\out{\hypertarget{method-Price-convert_to_commodity}{}}}
\subsection{Method \code{convert_to_commodity()}}{
Convert an amount from currency to commodity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$convert_to_commodity(currency_amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{currency_amount}}{Amount in currency units}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Price-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Convert to data frame row
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-print"></a>}}
\if{latex}{\out{\hypertarget{method-Price-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Price-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Price-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Price$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
