% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{PriceDB}
\alias{PriceDB}
\title{PriceDB R6 Class}
\description{
PriceDB R6 Class

PriceDB R6 Class
}
\details{
R6 class for managing a collection of prices.
Provides efficient lookup of prices by commodity and date.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PriceDB-new}{\code{PriceDB$new()}}
\item \href{#method-PriceDB-add}{\code{PriceDB$add()}}
\item \href{#method-PriceDB-get_prices}{\code{PriceDB$get_prices()}}
\item \href{#method-PriceDB-get_price}{\code{PriceDB$get_price()}}
\item \href{#method-PriceDB-get_rate}{\code{PriceDB$get_rate()}}
\item \href{#method-PriceDB-convert}{\code{PriceDB$convert()}}
\item \href{#method-PriceDB-as_tibble}{\code{PriceDB$as_tibble()}}
\item \href{#method-PriceDB-count}{\code{PriceDB$count()}}
\item \href{#method-PriceDB-print}{\code{PriceDB$print()}}
\item \href{#method-PriceDB-clone}{\code{PriceDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-new"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-new}{}}}
\subsection{Method \code{new()}}{
Create a new PriceDB
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$new(prices = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prices}}{Optional initial list of Price objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-add"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-add}{}}}
\subsection{Method \code{add()}}{
Add a price to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$add(price)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{price}}{A Price object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-get_prices"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-get_prices}{}}}
\subsection{Method \code{get_prices()}}{
Get all prices for a commodity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$get_prices(
  commodity_guid,
  currency_guid = NULL,
  start_date = NULL,
  end_date = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commodity_guid}}{GUID of the commodity}

\item{\code{currency_guid}}{Optional: filter by currency}

\item{\code{start_date}}{Optional: start of date range}

\item{\code{end_date}}{Optional: end of date range}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-get_price"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-get_price}{}}}
\subsection{Method \code{get_price()}}{
Get the most recent price for a commodity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$get_price(commodity_guid, currency_guid, as_of = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commodity_guid}}{GUID of the commodity}

\item{\code{currency_guid}}{GUID of the currency}

\item{\code{as_of}}{Optional: get price as of specific date}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-get_rate"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-get_rate}{}}}
\subsection{Method \code{get_rate()}}{
Get exchange rate between two commodities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$get_rate(from_guid, to_guid, as_of = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from_guid}}{Source commodity GUID}

\item{\code{to_guid}}{Target commodity GUID}

\item{\code{as_of}}{Optional: date for rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-convert"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-convert}{}}}
\subsection{Method \code{convert()}}{
Convert amount between commodities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$convert(amount, from_guid, to_guid, as_of = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount to convert}

\item{\code{from_guid}}{Source commodity GUID}

\item{\code{to_guid}}{Target commodity GUID}

\item{\code{as_of}}{Optional: date for rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Get all prices as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-count"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-count}{}}}
\subsection{Method \code{count()}}{
Get number of prices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-print"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PriceDB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PriceDB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PriceDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
