% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{pricedb_from_df}
\alias{pricedb_from_df}
\title{Create PriceDB from Data Frame}
\usage{
pricedb_from_df(df)
}
\arguments{
\item{df}{Data frame with price columns}
}
\value{
A PriceDB object
}
\description{
Build a PriceDB from a tibble of price data.
}
