% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{prices_from_df}
\alias{prices_from_df}
\title{Convert Data Frame to Price Objects}
\usage{
prices_from_df(df)
}
\arguments{
\item{df}{Data frame with price columns}
}
\value{
Named list of Price objects (keyed by guid)
}
\description{
Convert a tibble of price data to a list of Price objects.
}
