% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-csv.R
\name{print.gnucashr_csv_import}
\alias{print.gnucashr_csv_import}
\title{Print Method for gnucashr CSV Import}
\usage{
\method{print}{gnucashr_csv_import}(x, ...)
}
\arguments{
\item{x}{A gnucashr_csv_import object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for gnucashr CSV Import
}
