% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-logger.R
\name{print.logged}
\alias{print.logged}
\title{Print Logged Object}
\usage{
\method{print}{logged}(x, ...)
}
\arguments{
\item{x}{A logged object}

\item{...}{Ignored}
}
\description{
Print Logged Object
}
