% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{print.result}
\alias{print.result}
\title{Print Result Object}
\usage{
\method{print}{result}(x, ...)
}
\arguments{
\item{x}{A result object}

\item{...}{Ignored}
}
\description{
Print Result Object
}
