% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-balance-sheet.R
\name{print_balance_sheet}
\alias{print_balance_sheet}
\title{Print Balance Sheet}
\usage{
print_balance_sheet(bs, ...)
}
\arguments{
\item{bs}{Balance sheet from balance_sheet()}

\item{...}{Additional arguments passed to format_balance_sheet()}
}
\description{
Print Balance Sheet
}
