% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-income-statement.R
\name{print_income_statement}
\alias{print_income_statement}
\title{Print Income Statement}
\usage{
print_income_statement(is, ...)
}
\arguments{
\item{is}{Income statement from income_statement()}

\item{...}{Additional arguments passed to format_income_statement()}
}
\description{
Print Income Statement
}
