% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{quarterly_budget}
\alias{quarterly_budget}
\title{Create a Quarterly Budget}
\usage{
quarterly_budget(
  name,
  year = lubridate::year(Sys.Date()),
  description = NA_character_
)
}
\arguments{
\item{name}{Budget name}

\item{year}{Budget year (default: current year)}

\item{description}{Optional description}
}
\value{
A Budget object
}
\description{
Helper function to create a budget with quarterly periods.
}
