% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{result_map_err}
\alias{result_map_err}
\title{Map Over Error}
\usage{
result_map_err(x, fn)
}
\arguments{
\item{x}{A result object}

\item{fn}{Function to apply to error}
}
\value{
New result with transformed error
}
\description{
Apply a function to the error if Err, pass through if Ok.
}
