% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{result_match}
\alias{result_match}
\title{Pattern Match on Result}
\usage{
result_match(x, ok_fn, err_fn)
}
\arguments{
\item{x}{A result object}

\item{ok_fn}{Function to call if Ok}

\item{err_fn}{Function to call if Err}
}
\value{
Result of the matched function
}
\description{
Handle both Ok and Err cases with separate functions.
}
