% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{safe_read_gnucash}
\alias{safe_read_gnucash}
\title{Safe Read GnuCash}
\usage{
safe_read_gnucash(path, read_only = TRUE)
}
\arguments{
\item{path}{Path to GnuCash file}

\item{read_only}{Read-only mode}
}
\value{
Result with GnuCashDB or error
}
\description{
Open a GnuCash file with error handling.
}
