% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled-transaction.R
\name{scheduled_from_df}
\alias{scheduled_from_df}
\title{Convert Data Frame to ScheduledTransaction Objects}
\usage{
scheduled_from_df(df)
}
\arguments{
\item{df}{Data frame with schedxactions columns}
}
\value{
Named list of ScheduledTransaction objects
}
\description{
Convert a tibble of scheduled transaction data to objects.
}
