% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled-transaction.R
\name{ScheduledTransactionManager}
\alias{ScheduledTransactionManager}
\title{ScheduledTransactionManager R6 Class}
\description{
ScheduledTransactionManager R6 Class

ScheduledTransactionManager R6 Class
}
\details{
R6 class for managing scheduled transactions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScheduledTransactionManager-new}{\code{ScheduledTransactionManager$new()}}
\item \href{#method-ScheduledTransactionManager-add}{\code{ScheduledTransactionManager$add()}}
\item \href{#method-ScheduledTransactionManager-all}{\code{ScheduledTransactionManager$all()}}
\item \href{#method-ScheduledTransactionManager-due}{\code{ScheduledTransactionManager$due()}}
\item \href{#method-ScheduledTransactionManager-upcoming}{\code{ScheduledTransactionManager$upcoming()}}
\item \href{#method-ScheduledTransactionManager-as_tibble}{\code{ScheduledTransactionManager$as_tibble()}}
\item \href{#method-ScheduledTransactionManager-forecast_cash_flow}{\code{ScheduledTransactionManager$forecast_cash_flow()}}
\item \href{#method-ScheduledTransactionManager-print}{\code{ScheduledTransactionManager$print()}}
\item \href{#method-ScheduledTransactionManager-clone}{\code{ScheduledTransactionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new ScheduledTransactionManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-add"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-add}{}}}
\subsection{Method \code{add()}}{
Add a scheduled transaction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$add(sx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sx}}{A ScheduledTransaction object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-all"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-all}{}}}
\subsection{Method \code{all()}}{
Get all scheduled transactions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$all(enabled_only = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled_only}}{Only return enabled transactions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-due"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-due}{}}}
\subsection{Method \code{due()}}{
Get due scheduled transactions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$due(as_of = Sys.Date())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{as_of}}{Date to check (default: today)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-upcoming"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-upcoming}{}}}
\subsection{Method \code{upcoming()}}{
Get upcoming scheduled transactions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$upcoming(days = 30L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{days}}{Number of days to look ahead}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Get all as tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-forecast_cash_flow"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-forecast_cash_flow}{}}}
\subsection{Method \code{forecast_cash_flow()}}{
Calculate cash flow forecast from scheduled transactions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$forecast_cash_flow(
  months = 12L,
  account_mapping = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{months}}{Number of months to forecast}

\item{\code{account_mapping}}{Named list mapping account GUIDs to categories}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-print"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScheduledTransactionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScheduledTransactionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduledTransactionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
