% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{try_result}
\alias{try_result}
\title{Try Expression as Result}
\usage{
try_result(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
Result with value or error
}
\description{
Execute an expression and capture errors as Err result.
}
