% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{unwrap}
\alias{unwrap}
\title{Unwrap Result Value}
\usage{
unwrap(x, default = NULL)
}
\arguments{
\item{x}{A result object}

\item{default}{Optional default value if error}
}
\value{
The unwrapped value
}
\description{
Get the value from an Ok result, or abort for Err.
}
