% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monad-result.R
\name{unwrap_err}
\alias{unwrap_err}
\title{Unwrap Error}
\usage{
unwrap_err(x)
}
\arguments{
\item{x}{A result object}
}
\value{
The error value
}
\description{
Get the error from an Err result, or abort for Ok.
}
