% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation.R
\name{validate_consolidation}
\alias{validate_consolidation}
\title{Validate Consolidation}
\usage{
validate_consolidation(consolidated_tb)
}
\arguments{
\item{consolidated_tb}{Consolidated trial balance}
}
\value{
List with validation results
}
\description{
Check that consolidated trial balance is in balance.
}
