% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduled-transaction.R
\name{weekly_scheduled}
\alias{weekly_scheduled}
\title{Create a Weekly Scheduled Transaction}
\usage{
weekly_scheduled(
  name,
  amount,
  debit_account_guid,
  credit_account_guid,
  start_date,
  every_n_weeks = 1L,
  auto_create = FALSE
)
}
\arguments{
\item{name}{Transaction description}

\item{amount}{Transaction amount}

\item{debit_account_guid}{Account to debit}

\item{credit_account_guid}{Account to credit}

\item{start_date}{First occurrence}

\item{every_n_weeks}{Recurrence (1 = every week, 2 = biweekly)}

\item{auto_create}{Auto-create transactions}
}
\value{
A ScheduledTransaction object
}
\description{
Helper function to create a weekly recurring transaction.
}
